feat: up command scaffolding
